Replace device features with device binaries
