+++ Cargo.toml
@@ -5,15 +5,18 @@ rust-version = "1.91"
 version      = "0.1.0"
 
 [[bin]]
-name = "epd_photoframe"
-path = "./src/bin/main.rs"
+name = "e1001"
+path = "./src/bin/e1001.rs"
+
+[[bin]]
+name = "e1002"
+path = "./src/bin/e1002.rs"
+
+[[bin]]
+name = "e1004"
+path = "./src/bin/e1004.rs"
 
 [features]
-# Select exactly one device. Each device selects its panel driver and pinout.
-e1001 = [] # Seeed reTerminal E1001 (7", GDEY075T7 panel, 800x480 4-level grayscale)
-e1002 = [] # Seeed reTerminal E1002 (7", GDEP073E01 panel, 800x480)
-e1004 = [] # Seeed reTerminal E1004 (13", T133A01 panel, 1200x1600)
-
 # PPK2 bench-measurement mode (E1004 only). On boot, parks the SY6974B
 # charger in HIZ so the system rail runs entirely from the BAT input,
 # regardless of USB-C state. Intended for a Nordic Power Profiler Kit II
